Large commit intended to prepare current work for PR. Cleaned up comments on rgw_sal.cc and rgw_main.cc, and removed excessive log statements on rgw_user.cc and rgw_admin.cc. Cleaned up all tracer files to have indentation and form closer to ceph standard.
Made all functions besides friend and operator manipulation functions in
rgw::sal::User and rgw::sal::Bucket so that the tracer never has to
store its own variables. Removal of the static function empty into
is/info_empty for Bucket and User neccessitated the alterations of
several other files to function with the new boolean logic.

Added a new config value WITH_RADOSGW_TRACER that can run in tandem with
other store configs. In rgw::sal::Store*
StoreManager::init)storage_provider the logic has been reworked so that
depending on the actual store's config value the store pointer will
point to the actual store, and then it will be wrapped with a
tracer driver.